requirements.txt
Flask==3.0.0
gunicorn==20.1.0
google-cloud-vision 
google-cloud-storage